planning-graph: group known facts by predicate
This speeds up the search for a Fact that fulfils the current
precondition in matchPrecondition(), since we only need to look at facts
with a matching predicate.

It also speed up the check if a fact gained by performing a task was
already processed or not.

For the keyrack instance, this change speeds up execution by a factor of
about 30. For the 144 sample instances, speedups of about 5-8x can be
observed.